requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
requests
pillow
openai
